4.23*: Blog List Expansion, step 11

+++ Part4/blogList/tests/blog_api.test.js
@@ -3,15 +3,24 @@ const { test, after, beforeEach, describe } = require('node:test')
 const mongoose = require('mongoose')
 const supertest = require('supertest')
 const app = require('../app')
+const bcrypt = require('bcrypt')
 const helper = require('./test_helper')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 const api = supertest(app)
 
 beforeEach(async () => {
+  await User.deleteMany({})
+  const passwordHash = await bcrypt.hash('sekret', 10)
+  const user = new User({ username: 'root', name: 'superUser', passwordHash })
+  await user.save()
+
   await Blog.deleteMany({})
   await Blog.insertMany(helper.initialBlogs)
 })
+
+
 
 describe('Getting blogs', () => {
     test('all blogs are returned with same amount of blogs', async () => {
@@ -37,38 +46,46 @@ describe('Getting blogs', () => {
 
 describe('Creating a blog', () => {
     test('a valid blog can be added', async () => {
-    const newBlog = {
-        title: "New Blog",
-        author: "New Blog Writer",
-        url: "https://newblog.com/",
-        likes: 77
-    }
+        const loginRes = await api.post('/api/login').send({ username: 'root', password: 'sekret'})
+        const token = loginRes._body.token
+        
+        const newBlog = {
+            title: "New Blog",
+            author: "New Blog Writer",
+            url: "https://newblog.com/",
+            likes: 77
+        }
 
-    await api
-        .post('/api/blogs')
-        .send(newBlog)
-        .expect(201)
-        .expect('Content-Type', /application\/json/)
+        await api
+            .post('/api/blogs')
+            .send(newBlog)
+            .set('Authorization',  `Bearer ${token}`)
+            .expect(201)
+            .expect('Content-Type', /application\/json/)
 
-    const blogsAtEnd = await helper.blogsInDb()
-    assert.strictEqual(blogsAtEnd.length, helper.initialBlogs.length + 1)
+        const blogsAtEnd = await helper.blogsInDb()
+        assert.strictEqual(blogsAtEnd.length, helper.initialBlogs.length + 1)
 
-    const {id, ...savedBlogWithoutID} = blogsAtEnd[helper.initialBlogs.length]
-    assert.deepStrictEqual(savedBlogWithoutID, newBlog)
+        const {id, user, ...savedBlogWithoutID} = blogsAtEnd[helper.initialBlogs.length]
+        assert.deepStrictEqual(savedBlogWithoutID, newBlog)
     })
 })
 
 describe('Validating parameters', () => {
     test('blog with no like -> likes=0', async () => {
+        const loginRes = await api.post('/api/login').send({ username: 'root', password: 'sekret'})
+        const token = loginRes._body.token
+
         const newBlog = {
-        title: "No Likes",
-        author: "No Likes Writer",
-        url: "https://nolikes.com/"
-    }
+            title: "No Likes",
+            author: "No Likes Writer",
+            url: "https://nolikes.com/"
+        }
 
     await api
         .post('/api/blogs')
         .send(newBlog)
+        .set('Authorization',  `Bearer ${token}`)
         .expect(201)
         .expect('Content-Type', /application\/json/)
 
@@ -77,53 +94,92 @@ describe('Validating parameters', () => {
     })
 
     test('blog with no title/url gets 400', async () => {
+        const loginRes = await api.post('/api/login').send({ username: 'root', password: 'sekret'})
+        const token = loginRes._body.token
+
         const noTitleBlog = {
-        author: "No Likes Writer",
-        url: "https://nolikes.com/",
-        likes: 33
-    }
-    const noURLBlog = {
-        title: "No Likes",
-        author: "No Likes Writer",
-        likes: 33
-    }
+            author: "No Likes Writer",
+            url: "https://nolikes.com/",
+            likes: 33
+        }
+        const noURLBlog = {
+            title: "No Likes",
+            author: "No Likes Writer",
+            likes: 33
+        }
 
-    await api
-        .post('/api/blogs')
-        .send(noTitleBlog)
-        .expect(400)
-        .expect('Content-Type', /application\/json/)
+        await api
+            .post('/api/blogs')
+            .send(noTitleBlog)
+            .set('Authorization',  `Bearer ${token}`)
+            .expect(400)
+            .expect('Content-Type', /application\/json/)
 
-    await api
-        .post('/api/blogs')
-        .send(noURLBlog)
-        .expect(400)
-        .expect('Content-Type', /application\/json/)
+        await api
+            .post('/api/blogs')
+            .send(noURLBlog)
+            .set('Authorization',  `Bearer ${token}`)
+            .expect(400)
+            .expect('Content-Type', /application\/json/)
     })
 })
 
 describe('Deletion of a blog', () => {
     test('succeeds with status code 204 if id is valid', async () => {
-      const blogsAtStart = await helper.blogsInDb()
-      const blogToDelete = blogsAtStart[0]
+      const loginRes = await api.post('/api/login').send({ username: 'root', password: 'sekret'})
+      const token = loginRes._body.token
 
-      await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204)
+      const newBlog = {
+            title: "New Blog",
+            author: "New Blog Writer",
+            url: "https://newblog.com/",
+            likes: 77
+        }
+
+       await api
+            .post('/api/blogs')
+            .send(newBlog)
+            .set('Authorization',  `Bearer ${token}`)
+            .expect(201)
+            .expect('Content-Type', /application\/json/)
+
+      const blogsAtStart = await helper.blogsInDb()
+      const blogToDelete = blogsAtStart[2]
+
+      await api.delete(`/api/blogs/${blogToDelete.id}`).set('Authorization',  `Bearer ${token}`).expect(204)
 
       const blogsAtEnd = await helper.blogsInDb()
 
       const contents = blogsAtEnd.map((blog) => blog.url)
       assert(!contents.includes(blogToDelete.url))
 
-      assert.strictEqual(blogsAtEnd.length, helper.initialBlogs.length - 1)
+      assert.strictEqual(blogsAtEnd.length, helper.initialBlogs.length)
     })
 })
 
 describe('Updating a blog', () => {
     test('succeeds with status code 200', async () => {
+        const loginRes = await api.post('/api/login').send({ username: 'root', password: 'sekret'})
+        const token = loginRes._body.token
+
+        const newBlog = {
+                title: "New Blog",
+                author: "New Blog Writer",
+                url: "https://newblog.com/",
+                likes: 77
+            }
+
+       await api
+            .post('/api/blogs')
+            .send(newBlog)
+            .set('Authorization',  `Bearer ${token}`)
+            .expect(201)
+            .expect('Content-Type', /application\/json/)
+
         const updatedBlog = {
             title: "Updated React patterns",
             author: "Updated Michael Chan",
-            url: "https://reactpatterns.com/",
+            url: "https://newblog.com/",
             likes: 777
         }
 
@@ -133,6 +189,7 @@ describe('Updating a blog', () => {
         await api
             .put(`/api/blogs/${selectedBlog[0].id}`)
             .send(updatedBlog)
+            .set('Authorization',  `Bearer ${token}`)
             .expect(200)
             .expect('Content-Type', /application\/json/)
     })
